Major update of the HLT GUI
The GUI should be able to reset the histograms
monitored on request. It has implication on
the histogram handling:
- 3 lists of histograms are needed
  + The original data
  + The last timestamp
  + The difference between data and timestamp
- Preference in plotting has always the diff
  between original and timestamp
- Once the user presses the reset button, the
  last original becomes the new time stamp
- When new data is coming, the subtraction is
  always done automatically in case a time
  stamp is set
- All lists are synchronized via shared_ptrs
  (needed to be developed internal as external
  access is not guaranteed)
- The GUI itself keeps a shared pointer of all
  histograms it currently uses in order to
  prevent data races
- As we have asynchronous processes for
  message passing, histograms handling and GUI
  the relevant classes are derived from a base
  class "Synchronized" implementing the
  locking

Furthermore the program is put under GPL V3.
Code from EmcalZmqHelpers however is  put under
the BSD License in order to fullfil other license
requirements. Authors are added again under the
doxygen style format. The indentation style is
defined with the standard K&R style (4 spaces).